requirements.txt
click==8.1.3
Pillow==9.2.0
progress==1.6
